hack/debug-agent/requirements.txt
pyjwt
requests
pyyaml
toml
